Make thE Gambler Game so user can get there win and loss percentage

+++ Practice_Question/square.py
@@ -1,3 +1,3 @@
-num=float(input("Enter the number : "))
+num=int(input("Enter the number : "))
 sqrt_num=num **0.5
 print(f"The square root of number is {sqrt_num}")
